restructured replacementMobys to improve scalability. disabling pending through location syncing
