+++ Lawrence/runtime/runtime.lua
@@ -22,16 +22,16 @@ require 'runtime.ui.InputElement'
 require 'HybridMoby'
 
 require 'Level'
-require 'runtime.levels.Novalis'
-require 'runtime.levels.Kerwan'
-require 'runtime.levels.Eudora'
-require 'runtime.levels.BlargStation'
-require 'runtime.levels.Rilgar'
-require 'runtime.levels.Umbris'
-require 'runtime.levels.Gaspar'
-require 'runtime.levels.Orxon'
-require 'runtime.levels.Pokitaru'
-require 'runtime.levels.DreksFleet'
+-- require 'runtime.levels.Novalis'
+-- require 'runtime.levels.Kerwan'
+-- require 'runtime.levels.Eudora'
+-- require 'runtime.levels.BlargStation'
+-- require 'runtime.levels.Rilgar'
+-- require 'runtime.levels.Umbris'
+-- require 'runtime.levels.Gaspar'
+-- require 'runtime.levels.Orxon'
+-- require 'runtime.levels.Pokitaru'
+-- require 'runtime.levels.DreksFleet'
 
 null = {}
